Try once again make in threads
It's still unclear why make time is faster now. There is not supposed to
be any cache. But lets change dockerfile and see if anything will be
affected.

Lets try 2 threads as we by default have 2 vCPUs.

Additionally,
2.5 minutes for install
6.5 minutes for make

If we anyway will create a special docker image for dlisio, makes sense
to create one here also.