Trigger bigendian setup creation job triggered by tag
If we actually decide to put images out there for speedup, we would need
an option to build them if something changes.
As practice showed, building s390x images on local architecture can be a
challenge.
So it might be possible to run the job on CirceCI and download/store
artifacts later.

For now try to create a job that's going to be triggered with special
tags.
However not sure that understand mechanism correctly.