Check remote docker execution
OK, previous version works nicely.

Now lets see if it would be even possible to use remote_docker instead
of the machine.
It's unclear if there would be any benefits to it.
It doesn't seem so. Still everything must happen in the dockerfiles, not
in config.yml because s390x things are available only inside remote
docker.

Caching seems to be the same in image/docker, so that's independent
step.